前端：前端博客详情，评论UI半完成

+++ blog_core/src/main/java/org/example/base/vo/BaseVO.java
@@ -27,11 +27,11 @@ public class BaseVO<T>  {
      * 当前页
      */
 
-    private Long currentPage;
+    private Integer currentPage;
 
     /**
      * 页大小
      */
 
-    private Long pageSize;
+    private Integer pageSize;
 }
